Site

+++ db_postgres/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Integer, String, ForeignKey, BigInteger
 from sqlalchemy.orm import relationship
 from db_postgres.db import Base
 
@@ -6,7 +6,7 @@ from db_postgres.db import Base
 class User(Base):
     __tablename__ = "users"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     username = Column(String, nullable=False)
     role = Column(Integer)
 
@@ -27,7 +27,7 @@ class PlanSpending(Base):
     __tablename__ = "plan_spending"
 
     id = Column(Integer, primary_key=True, autoincrement=True)
-    user_id = Column(Integer, ForeignKey("users.id", ondelete="CASCADE"))
+    user_id = Column(BigInteger, ForeignKey("users.id", ondelete="CASCADE"))
     category = Column(String, nullable=False)
     amount_money = Column(Integer, nullable=False)
 
@@ -38,7 +38,7 @@ class StatusScheduler(Base):
     __tablename__ = "status_scheduler"
 
     id = Column(Integer, primary_key=True, autoincrement=True)
-    user_id = Column(Integer, ForeignKey("users.id", ondelete="CASCADE"))
+    user_id = Column(BigInteger, ForeignKey("users.id", ondelete="CASCADE"))
     status_notification = Column(Integer)
     status_update = Column(Integer)
 
@@ -56,7 +56,7 @@ class Expense(Base):
     __tablename__ = "expenses"
 
     id = Column(Integer, primary_key=True, autoincrement=True)
-    user_id = Column(Integer, ForeignKey("users.id", ondelete="CASCADE"))
+    user_id = Column(BigInteger, ForeignKey("users.id", ondelete="CASCADE"))
     category = Column(String, nullable=False)
     amount_expenses = Column(Integer, nullable=False)
 
@@ -67,8 +67,8 @@ class Share(Base):
     __tablename__ = "shares"
 
     id = Column(Integer, primary_key=True, autoincrement=True)
-    master_id = Column(Integer, ForeignKey("users.id", ondelete="CASCADE"))
-    slave_id = Column(Integer, ForeignKey("users.id", ondelete="CASCADE"))
+    master_id = Column(BigInteger, ForeignKey("users.id", ondelete="CASCADE"))
+    slave_id = Column(BigInteger, ForeignKey("users.id", ondelete="CASCADE"))
 
     master = relationship("User", back_populates="shares_as_master", foreign_keys=[master_id])
     slave = relationship("User", back_populates="shares_as_slave", foreign_keys=[slave_id])
@@ -77,7 +77,7 @@ class Reminders(Base):
     __tablename__ = "reminders"
     
     id = Column(Integer, autoincrement=True, primary_key=True)
-    user_id = Column(Integer, ForeignKey("users.id", ondelete="CASCADE"))
+    user_id = Column(BigInteger, ForeignKey("users.id", ondelete="CASCADE"))
     status = Column(Integer)
     
     user = relationship("User", back_populates="reminders")
@@ -86,7 +86,7 @@ class EveryWaste(Base):
     __tablename__ = "every_waste"
     
     id = Column(Integer, primary_key=True, autoincrement=True)
-    user_id = Column(Integer, ForeignKey('users.id', ondelete="CASCADE"))
+    user_id = Column(BigInteger, ForeignKey('users.id', ondelete="CASCADE"))
     category = Column(String, nullable=False)
     expense = Column(Integer, nullable=False)
     data_time = Column(String)
@@ -98,14 +98,14 @@ class Wishes(Base):
     __tablename__ = "wishes"
     
     id = Column(Integer, autoincrement=True, primary_key=True)
-    user_id = Column(Integer)
+    user_id = Column(BigInteger)
     comment = Column(String)
     
 class MonthlyReport(Base):
     __tablename__ = "monthly_report"
     
     id = Column(Integer, primary_key=True, autoincrement=True)
-    user_id = Column(Integer, ForeignKey('users.id', ondelete="CASCADE"))
+    user_id = Column(BigInteger, ForeignKey('users.id', ondelete="CASCADE"))
     path_to_json = Column(String)
     
     user = relationship("User", back_populates="monthly_report")
